CI: Improve availability.rb check
Improves ci/availability.rb to check whether a test case file with
expected name and extension exists for each rule.

To make the current rules pass, it was necessary to create placeholder
test case files for rules that didn't have one. These will be filled
out with actual example code as unit tests are implemented.

The check is also prepared to check if test case files are empty, which
is turned off for the time being, but when all files contain code, it
should be enabled.